test(frontend): update template validators tests to match simplified implementation
- Remove emptyContentValidator tests (function no longer exists)
- Remove MockBudgetTemplatesState complexity (replaced with simple string array)
- Update duplicateNameValidator tests to use synchronous validation
- Remove EMPTY_CONTENT error references
- Fix import paths and maintain comprehensive test coverage

All 13 tests now pass with the simplified validator implementation.
