+++ frontend/projects/webapp/src/app/feature/budget-templates/create/components/create-template-form.ts
@@ -1,0 +1,372 @@
+import {
+  ChangeDetectionStrategy,
+  Component,
+  computed,
+  effect,
+  inject,
+  input,
+  output,
+} from '@angular/core';
+import { toSignal } from '@angular/core/rxjs-interop';
+import { JsonPipe } from '@angular/common';
+import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatDividerModule } from '@angular/material/divider';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { type BudgetTemplateCreate } from '@pulpe/shared';
+import { BudgetTemplatesState } from '../../services/budget-templates-state';
+import { DefaultWarningPanelComponent } from '../ui/default-warning-panel.component';
+import {
+  duplicateNameValidator,
+  FORM_LIMITS,
+  getTemplateValidationErrorMessage,
+  VALIDATION_MESSAGES,
+} from './template-validators';
+
+@Component({
+  selector: 'pulpe-create-template-form',
+  imports: [
+    ReactiveFormsModule,
+    MatButtonModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatCheckboxModule,
+    MatIconModule,
+    MatProgressSpinnerModule,
+    MatTooltipModule,
+    MatDividerModule,
+    DefaultWarningPanelComponent,
+    JsonPipe,
+  ],
+  template: `
+    @if (state.isLoading() && state.templatesData.status() === 'loading') {
+      <div class="flex justify-center items-center py-16">
+        <mat-spinner diameter="40"></mat-spinner>
+      </div>
+    } @else {
+      <div
+        class="relative space-y-4 md:space-y-6"
+        data-testid="template-form-container"
+      >
+        <!-- Loading overlay during form submission -->
+        @if (isCreating()) {
+          <div
+            class="absolute inset-0 bg-surface/50 flex items-center justify-center z-10 rounded-corner-medium"
+          >
+            <div class="flex flex-col items-center gap-3">
+              <mat-spinner diameter="32"></mat-spinner>
+              <span
+                class="text-body-medium text-on-surface-variant font-medium"
+              >
+                Création en cours...
+              </span>
+            </div>
+          </div>
+        }
+
+        @if (state.templateCount() > 0) {
+          <div class="text-body-medium text-on-surface-variant">
+            {{ state.templateCount() }}/{{ state.MAX_TEMPLATES }} modèles créés
+          </div>
+        }
+
+        <form
+          [formGroup]="templateForm"
+          (ngSubmit)="onSubmit()"
+          (keydown.escape)="onEscapeKey()"
+          data-testid="template-form"
+        >
+          <!-- Grid container with responsive columns and MD3 spacing -->
+          <div class="flex flex-col gap-4 md:gap-6 pt-4">
+            <!-- Name field - full width for optimal UX -->
+            <mat-form-field appearance="outline" class="w-full">
+              <mat-label>Nom du modèle</mat-label>
+              <input
+                matInput
+                formControlName="name"
+                required
+                [maxlength]="formLimits.NAME_MAX_LENGTH"
+                data-testid="template-name-input"
+              />
+              <mat-hint align="end">
+                {{ templateForm.get('name')?.value?.length }}/{{
+                  formLimits.NAME_MAX_LENGTH
+                }}
+              </mat-hint>
+              @if (nameValidationError()) {
+                <mat-error data-testid="name-error" id="name-error-message">
+                  {{ nameValidationError() }}
+                </mat-error>
+              }
+            </mat-form-field>
+
+            <mat-form-field appearance="outline" class="w-full">
+              <mat-label>Description (optionnelle)</mat-label>
+              <textarea
+                matInput
+                formControlName="description"
+                rows="3"
+                aria-label="Description du modèle (optionnel)"
+                data-testid="template-description-input"
+              ></textarea>
+              <mat-hint align="end">
+                {{ templateForm.get('description')?.value?.length }}/{{
+                  formLimits.DESCRIPTION_MAX_LENGTH
+                }}
+              </mat-hint>
+              @if (templateForm.get('description')?.errors) {
+                <mat-error>
+                  @if (templateForm.get('description')?.errors?.['maxlength']) {
+                    La description ne doit pas dépasser
+                    {{ formLimits.DESCRIPTION_MAX_LENGTH }} caractères.
+                  }
+                </mat-error>
+              }
+            </mat-form-field>
+
+            <div class="flex items-center gap-2">
+              <mat-checkbox
+                formControlName="isDefault"
+                data-testid="template-default-checkbox"
+                aria-label="Définir comme modèle par défaut"
+                class="min-h-[44px] flex items-center"
+              >
+                Modèle par défaut
+              </mat-checkbox>
+              <mat-icon
+                matTooltip="Le modèle par défaut est sélectionné par défaut pour créer de nouveaux budgets mensuels."
+                matTooltipPosition="above"
+                aria-label="Information sur le modèle par défaut"
+                class="!text-on-surface-variant cursor-help"
+                tabindex="0"
+              >
+                info
+              </mat-icon>
+            </div>
+
+            <!-- Default warning panel -->
+            @if (showDefaultWarning()) {
+              <div class="col-span-1">
+                <pulpe-default-warning-panel
+                  id="default-warning"
+                  [message]="warningMessage()"
+                />
+              </div>
+            }
+
+            <!-- Enhanced error message section -->
+            @if (state.businessError()) {
+              <div class="col-span-1">
+                <div
+                  class="p-4 bg-error-container text-on-error-container rounded-corner-medium flex items-start gap-3"
+                  role="alert"
+                  aria-live="assertive"
+                >
+                  <mat-icon class="text-on-error-container flex-shrink-0 mt-0.5"
+                    >error</mat-icon
+                  >
+                  <span class="text-body-medium leading-relaxed">
+                    {{ state.businessError() }}
+                  </span>
+                </div>
+              </div>
+            }
+
+            <!-- Template limit information -->
+            @if (!state.canCreateMore()) {
+              <div class="col-span-1">
+                <div
+                  id="limit-reached-info"
+                  class="p-4 bg-error-container text-on-error-container rounded-corner-medium flex items-start gap-3"
+                  role="status"
+                  aria-live="polite"
+                >
+                  <mat-icon class="text-on-error-container flex-shrink-0 mt-0.5"
+                    >warning</mat-icon
+                  >
+                  <span class="text-body-medium leading-relaxed">
+                    Limite de {{ state.MAX_TEMPLATES }} modèles atteinte.
+                    Supprimez un modèle existant pour en créer un nouveau.
+                  </span>
+                </div>
+              </div>
+            }
+          </div>
+
+          <!-- Divider to separate form from actions -->
+          <mat-divider class="!mt-4 !mb-4" />
+
+          <!-- Action buttons aligned right with mobile-first responsive layout -->
+          <div class="flex flex-col-reverse gap-3 sm:flex-row sm:justify-end">
+            <button
+              matButton
+              (click)="cancelForm.emit()"
+              [disabled]="isFormDisabled()"
+              data-testid="cancel-button"
+              type="button"
+            >
+              Annuler
+            </button>
+            <button
+              matButton="filled"
+              type="submit"
+              [disabled]="!isFormValidForSubmission()"
+              data-testid="submit-button"
+            >
+              {{ submitButtonText() }}
+            </button>
+          </div>
+        </form>
+      </div>
+    }
+  `,
+  styles: `
+    :host {
+      display: block;
+    }
+  `,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class CreateTemplateForm {
+  // Injected dependencies
+  #fb = inject(FormBuilder);
+  protected state = inject(BudgetTemplatesState);
+  protected formLimits = FORM_LIMITS;
+
+  isCreating = input(false);
+
+  addTemplate = output<BudgetTemplateCreate>();
+  cancelForm = output<void>();
+  formReset = output<void>();
+
+  templateForm = this.#fb.group({
+    name: [
+      '',
+      {
+        validators: [
+          Validators.required,
+          Validators.maxLength(FORM_LIMITS.NAME_MAX_LENGTH),
+          duplicateNameValidator(
+            this.state.templatesData.value()?.map((t) => t.name.toLowerCase()),
+          ),
+        ],
+      },
+    ],
+    description: [
+      '',
+      [Validators.maxLength(FORM_LIMITS.DESCRIPTION_MAX_LENGTH)],
+    ],
+    isDefault: [false],
+  });
+
+  formValues = toSignal(this.templateForm.valueChanges, {
+    initialValue: this.templateForm.getRawValue(),
+  });
+  formStatus = toSignal(this.templateForm.statusChanges, {
+    initialValue: this.templateForm.status,
+  });
+  isFormValid = computed(() => this.formStatus() === 'VALID');
+
+  // Computed signals reading from consolidated formValues
+  isDefaultChecked = computed(() => this.formValues()?.isDefault ?? false);
+
+  showDefaultWarning = computed(() => {
+    const hasDefault = this.state.hasDefaultTemplate();
+    const isDefaultChecked = this.isDefaultChecked();
+    return hasDefault && isDefaultChecked;
+  });
+
+  warningMessage = computed(() => {
+    const defaultTemplate = this.state.currentDefaultTemplate();
+    return defaultTemplate?.name
+      ? `Le modèle "${defaultTemplate.name}" ne sera plus le modèle par défaut.`
+      : 'Un modèle par défaut existe déjà. Il sera remplacé.';
+  });
+
+  nameValidationError = computed(() => {
+    const isFormValid = this.isFormValid();
+
+    if (isFormValid) return null;
+    return getTemplateValidationErrorMessage(
+      this.templateForm.get('name')?.errors ?? null,
+      'nom',
+    );
+  });
+
+  isFormDisabled = computed(() => this.isCreating());
+
+  // Smart submit button state management
+  submitButtonText = computed(() => {
+    const isCreating = this.isCreating();
+    const canCreateMore = this.state.canCreateMore();
+    if (isCreating) return 'Création...';
+    if (!canCreateMore) return 'Limite atteinte';
+    return 'Créer';
+  });
+
+  isFormValidForSubmission = computed(() => {
+    const isFormValid = this.isFormValid();
+    const isPending = this.isCreating();
+    const canCreateMore = this.state.canCreateMore();
+    return isFormValid && !isPending && canCreateMore;
+  });
+
+  constructor() {
+    // Effect to disable/enable form fields based on isCreating signal
+    effect(() => {
+      if (this.isCreating()) {
+        this.templateForm.disable();
+      } else {
+        this.templateForm.enable();
+      }
+    });
+  }
+
+  onSubmit() {
+    this.state.businessError.set(null);
+    this.templateForm.markAllAsTouched();
+    if (!this.isFormValidForSubmission()) {
+      if (!this.state.canCreateMore()) {
+        this.state.businessError.set(
+          VALIDATION_MESSAGES.TEMPLATE_LIMIT_REACHED(this.state.MAX_TEMPLATES),
+        );
+      }
+      return;
+    }
+
+    // Form is valid, create template from form values
+    const formVal = this.formValues();
+    const template: BudgetTemplateCreate = {
+      name: formVal?.name?.trim() ?? '',
+      description: formVal?.description?.trim() || undefined,
+      isDefault: formVal?.isDefault ?? false,
+      lines: [],
+    };
+
+    this.addTemplate.emit(template);
+    // Reset form immediately after successful submission
+    this.resetForm();
+  }
+
+  // Enhanced form reset with better UX
+  resetForm() {
+    this.templateForm.reset();
+    this.templateForm.markAsUntouched();
+    this.templateForm.markAsPristine();
+    this.state.businessError.set(null);
+
+    this.formReset.emit();
+  }
+
+  // Enhanced keyboard navigation
+  onEscapeKey() {
+    if (!this.isCreating()) {
+      this.cancelForm.emit();
+    }
+  }
+}
